Player can now only swap once every 2 frames. Swaps occur 1 frame after their input is received. Panels that have just begun hovering may match, but may not chain. Together, these changes make situations involving swapping a panel into a matching position above panels that are about to clear behave correctly.
That is:
It is impossible to make a horizontal 3x1 chain this way.
It is impossible to make a vertical 3x1 chain by swapping the topmost
panel into position.
For the other two panels, the following is true.
One frame, if you attempt the swap, the panels will match prematurely.
The next 3 frames, if you swap, a gap (of size 1-3) will form in the
falling panels and you will get a chain.
The next frame, you will be unable to swap.