2P given red frame
Frames and walls are now contained in <theme>/frame, and are numbered uniquely. 2P now has the red frame from the original SNES games, and frames for 3P to 8P have been added for the eventual addition of increased max players.

Special thanks to TheWolfBunny on DeviantArt for 3P-8P frame and wall palettes.