replaced TA character garbage
previously, TA characters mostly used their PdP counterpart's garbage.
Thanks for the assets @Miz

Next, we need a way to know whether to overwrite characters that have already been copied to the save directory.
Maybe we put in a file like \default_data\version.txt, and if the number in it is higher than the version in appdata, we overwrite the files.  Something like that.