Game Over Enhancements
This pull requests implements several improvements to the game over transition including:
Stacks don't dissappear immediately
Players must wait a short period before exiting to avoid accidently skips when a swap was intended.
The stack switches to the "shocked" face for the panels to help indicate who lost.
The timer stops showing the final time.
A pop effect is added to the top of the losing player (someday should be switched to a smoke effect)
The cursors go away once the game is over.
"Press any key" is shown on screen.

More improvements can be done, but this gets us started and the patch was already getting too big.

In order to pull this off, the biggest change that needed to be made was allow parts of the game to run after game over
to allow special effects, update the panels, etc.

To do this I have refactored and cleaned up the logic for creating, running, and ending local and foriegn games.
Including:
    Game over state is tracked in a global including the exact frame the first player lost
    The Stack.run method is consolidated with a variable is_local to indicate a local vs foriegn run.
    Logic updated for which portions of the run logic run when the game is over
    Game stopwatch (the one used for the visual timer) doesn't increment after game over
    Time left on time attack clock can't go negative (plus bullet proofing)
    Base panels "shocked" image updated to be the same as the dim image, otherwise the screen looks empty on gameover