Spectator room init fix
From testing, a way to get the room to fail to init for a spectator is to have the match start before the spectator leaves the game over screen. This fixes that by ensuring the spectator goes back to the room without the game over screen. We will have to figure out a way in the future so that the spectator can also see the game over screen, but this is the best solution for now that pretty much removes the init failures.