Fix game init failures Issue #227
The order the messages come through from the server is very important. Also if we drop some of those messages, we can get in a corrupted state.
To fix this, we create a queue for the other time sensitive data messages. (Panel buffers for now, we probably should do input as well.)
We only process that queue when we are in a game state, otherwise we might eat the inputs for the next game before we are setup.
Also added vscode files so debugging can work.