Patch: Spectator Room Bug
As player Dekrates put it, this bug was triggered by the following:
- two players join the server and spectate the same match
- these two players then decide to stop spectating and play vs each other
- they will have the same rating points as the players they just spectated
- after the first match of them finishes everyone gets thrown out of their matches/lobbies
- can be circumvented by going back to menu before starting the match that would cause the crash

The problem was that the server only would increment the room number if the players were not in a room. However, when leaving spectator, the game did not clear that room status from the spectators. Therefore, the game believed the players were already in a room and did not increment the room number. This overwrote the previous room, leaving those players in a room that no longer existed.