Log Dropped Frames, create render method on match
When more than one game loop is run in a row the following negative side effects happen:
1. If this is a local game, the player can't input between those frame simulations.
2. The game doesn't render the first frame.

We need to avoid doing more than 1/60th of a second of work in the loop or we will hit this.

To start, lets log when this happens and display it onscreen when it happens a lot so we can diagnose how common this is.