Fix going away
In the previous fix I was worried about turning the music back on before stopping it, but its safe to do since we immediately stop it before the engine does anything.

The bug was the music was removed from the global list before it was set back to normal percentage.