Better MultiBar Stop Time Indication
The Stop Time Bars stack on top of each other to indicate remaining stop time. However, the bars have the potential to have more than one shrink at a time, causing... "visual jank". This change makes it so that this will not be the case by displaying only the prestop + stop bar together, or only the shake bar alone if the shake frames are greater than or equal to the sum of the prestop and stop frames.