Placement Match Limit Change
The survey on a change from 50 matches to 25 was performed on the Panel Attack Discord Server. The final vote was:

Yes: 25
No: 6*

The vote ended in a roughly 5:1 ratio in favor of changing the amount of matches to 25. However, to give the 6 members who voted no a part in this, I took the percentile of the no votes (which was about 20%), multiplied it by 25 (the amount of matches we were reducing) and ended up with 5, so I added 5 more matches to the 25 matches to compromise.