Default VSync to off
It seems with various OSs + gpus + love we can't guarantee love.update to be called at 60fps always with v sync turned on, but for some reason it seems to be more consistent with vsync turned off.

This will cause missed inputs whenever it happens, and isn't very obvious to the user what's going on.

Therefore we should default vsync to off. The opposite problem IS obvious. The screen refreshing in a weird way. If that bothers the user, it will be much more obvious what to do, go turn vsync on.