Reduce log output on client (#342)
Introduces a logger with customizable verbosity and timestamps and cleans up several files with noisy logs on the client.

There are still several active print statements in the codebase (my editor shows 175), so more cleanup work can be done, but I addressed the noisiest log statements that showed up in my own casual playing and spectating. The majority of the remaining print statements (120 of them, oof) are in server.lua, so that should be addressed eventually.

This partially addresses #341.