Fix rollback off by 1 errors
Each interaction of the stack with it's opponent needs to happen on the frame after or later so the order they run doesn't matter.
Thus chain ending and attacks should happen on the frame after they happened.
We also only need to rollback to the frame AFTER what happened, because thats the first frame things can be affected.
Also did some other cleanups to make debugging easier.