Fix Desyncs when stoppers are popped and extended on the same frame
If a stopper is extended by one player and popped by the other player on the same frame, the order of which stack runs first matters.
To avoid this, don't pop garbage off the telegraph until both players have extended stoppers.
That data isn't processed till next frame anyways, so its pretty convenient to move this.