Fix rollback breaks (#463)
I lost my way and forgot the golden rule:
Interactions with the other player need to always happen on future frames, never the current frame.

Thus I tried to have multiple phases of running and confuse things. One player might cause the other player to rollback in the midst of the phases, thus causing the CLOCK to increment at the wrong time.

Reverted the multiple phase stuff.

Instead just make sure attacks, extending attacks, and ending chains all happen on the frame after they happen. Then when we rollback to the frame after they happen, we are guaranteed they happened regardless of which side ran first.

Most of these changes are just the rename from frame_earned to timeAttackInteracts to clarify the +1.
The important changes are in telegraph:push and telegraph:chainEnded as well as the revert of the engine changes.