lag improvements
Don't crash the game when lag goes too far to continue the match.
Instead as soon as one player gets too far behind, record the replay, present an error, and return to the lobby.
Spectators will get a message as well along with information on the average lag for both players.
Show a bug icon when games are close to quitting for spectators
Also some debug cleanups and stoner mode improvements