Implement love.run for better performance
Garbage collection was causing major spikes in cpu which reduced performance and contributed to input lag.

To help with this we implement our own love.run similar to the default.

What is different is we only update / render at 60fps and we garbage collect on the other updates. This means we have less garbage build up which reduces spikes.

Other changes include
Debug options menu for inducing lag in replays and vs online
Performance logger for tracking this in the future
Vsync is now always off to support this feature