Cleanup rollback memory
The telegraph and garbage queue were holding onto the stack class in there rollback copies causing a whole swatch of memory to stay around till the next game.

To prevent this, we don't store references to the stack in the rollback save states, and hook it back up on rollback.