Music test revamp (#490)
* Music Test overhaul

options.lua (specifically the music room bit)
-It's a click menu now
-Added support for character bundles and dynamic music
-Forced garbage collection while loading to prevent out of RAM crash
-You can play a song in the music room, leave the music room, and keep that song playing until you go to another game mode/select screen/reload the music room (does not work for dynamic music, main theme is restored as fallback)
-Added character sound effects so this is now more of a Sound Test than a Music Room. Characters without songs are now loaded as a result.

character.lua changes reduces loading the same file when sounds fallback to base combo/chain. Also needed to allow sounds to be compared to see if this fallback was used in Music Test.

mainloop.lua
main theme put on hold if music is already playing brought over from options from the music room.

* Localization bits, visual bug fix

Fixed visual bug with sound type when leaving the music room on an entry with valid music, and reloading with the first entry not having music.