Title Screen and scrolling background support
Made a new TilingImage class that supports tiling images that optionally scroll
Made all the backgrounds use tiling image
Added speeds to the theme to allow scrolling of the backgrounds
Added a optional title background to the theme. When set, it will be the first screen 
The menu will also be shifted up as its assumed there is no logo on the main menu
If not set, nothing should change.
Loading strings are implemented better
Removed references to “legacy canvas size” as it doesn’t matter anymore
Reworked the math
Fixed gprint and gprintf to set the color back to white after changing it
Moved main_menu_screen_pos to the theme
Make sure the lobby position isn’t too high so the welcome text is still visible
Resized the base backgrounds as they weren’t the right resolution
Added the copper league image
Fixed the size of the bronze league image