Fix loading in stages with no background
When a stage doesn't have a background, it would use the default background, since we already made an UpdatingImage there, it would try to construct with one which was invalid. This made us get stuck loading the stage.

Instead, just make un UpdatingImage when we finally set the background do the stage.