Config on open
We want to load all our config files before love startup so we can position the window correctly the first time, rather than have it jump.

Move around the code so that conf.lua can already have the values it needs.

Also add an option for picking the scale factor of the canvas. Auto is the default which picks the best incremental scale factor. Scale can be used to always use all the window real estate, finally you can force to a particular scale.

Made the game start maximized on first launch
Now storing and restoring the width, height, and maximized state of the game

Also cleaned up the order of the config file and removed unused parameters in modern love.