Fix top performance problems
I profiled and found the engine was spending a majority of the time saving rollback info and calculating the exclude matches test.

If we are behind the time that the opponent's new attacks would land, then we don't need to rollback
So we don't save the rollback info for performance reasons

To help speed up exclude matches, I saved off the result for each panel once, so we don't need to recalculate it. Note we still call it manually later on because the state may have changed changing the result.

After all these changes performance is generally somewhere from 2x to 6x faster on my machine.