Fix blurry graphics (#648)
To make sure graphics are not blurry and crisp in all scales we have the following requirements:

1. The global canvas needs to have scale * canvas size pixels otherwise it will not be able to hold all the details for assets of equal or higher resolution.
2. The high res assets need to scale down linearly for 1.5 scale so they don't look jaggy
3. 1x assets need to be upscaled using "nearest" if they are intended to be blocky. (most TA graphics)
4. 1x assets need to be upscaled using "linear" if they are not intended to be block. ("HD" Panels like Miz's)
The problem is there is no way to know if we are in case 3 or 4 for the most part, so what we should do is assume 3 by default for 1x images. However panels that are 48 tall, we will default to 4, because they probably want to be linear

In the future we can make a way to override the default for particular images.

We were creating canvases with non standard scales like "0.8654" when you were at odd zoom levels.
This creates all kinds of bad artifacts.
We were also creating fonts with bad scales.
Instead we want to snap DPI to reasonable boundaries. 1, 1.5, 2 etc

This makes almost all graphics look good, but we take special care with the 1x panels to make sure the "HD" ones use linear upscaling