Fix number and time labels
number and time labels had their scale and offsets hard coded, which meant we could esaily change assets or use higher resolution ones

they also had bugs with numbers and letters overlapping, especially when we drew a number with the time asset

instead we use a consistent pixel font algorithm for all of them which supports left center and right alignment

This means we have to fix all the theme positions and scales, but at least they are consistent now.

Updated the base theme to look good.