improve behavior of setting portrait mode
on desktop, window will now keep its height, and either decrease or increase the window width.  Window will now be resizable as normal.

portraitMode will now be saved to config.

portraitMode button on the main menu should now also show whether portait mode is on or off, but I am still running into an issue with main menu not updating this text, or background image.  If you go to another menu, and back again, it reflects the change.