Dont save an entire replays rollback
Rollback currently uses roughly 4 MB per second of memory. If we store that much for an entire game, it can quickly balloon out of control, bringing the garbage collector to its knees and potentially crashing the game.

Until we have a better fix, I have changed it to only save the same amount we save for worse case rollback during live matches which is about 4 seconds.

This means for now we can only rewind 4 seconds, but that will have to do for now until we can optimize this scenario.