Update auto updater with version and ability to override run (#796)
We discovered that when running from the autoupdater, run is already loaded and executing in a while loop from love2d, and so we can't override it or change it.

What we can do however is override or change methods that that run loop executes.

the simplest thing to do is just override love.timer. I tested and that works, but its not clean or as well tested.

This proves another thing we can do, but will require an autoupdater update.

If we provide a run implementation that has our own internal call to run, we can change it out however we want whenever we want. This make things a lot more modular, which should allow us to change / fix things again without ever needing to rev the autoupdater again.

I have also added a version number to the autoupdater and a warning recommending updating if you are out of date.