Check matches refactor (#898)
An extraction of #855 that isn't polluted by #854 (and should therefore not have any significant conflicts with #891)

Cleans up check_matches and makes it much faster while at it.
The speed is mostly gained by greatly reducing the amount of panels checked via using the new panel.stateChanged flag and optimizing the scope of the flood queue for finding cleared garbage.

This PR already includes the fixes #888 and #896.

Summary

Basically the entire content of the previous check_matches function has moved to the new file engine/checkMatches.lua.
The new checkMatches function is comprised of many high level functions and more or less fits on one monitor height worth of space.

Changes outside of checkMatches.lua

SFX

The only part of the old check_matches that made it outside is cancelling the garbage land SFX if a combo or chain was achieved on the same frame the garbage land would have played. Previously at the very end of check_matches it is now getting cancelled in the sound section of Stack.simulate

New Row

In new_row, panels that ascended to row 1 are now explicitly flagged as stateChanged so that they're eligible for checkMatches. This is commented accordingly.

Panel.update

Panel.update now routinely clears 3 flags that are used by checkMatches on every single frame.
Additionally there is a new Panel function for matching non-garbage panels.

Tests

There's a new basicVsTest2 that covers 2 cases of desync that were reported to me by chaos and/or Musichael (can't quite remember)

Other changes

Name changes

As they got usually only called in their negated form, exclude_match and exclude_swap had their meaning and return values inversed into canMatch and canSwap. Calls were adjusted in the respective locations.

Debug

For debugging purposes, debug now shows debug lines for panels up to row 13 if it exists. I added this to troubleshoot the issue covered by basicVsTest2 that matched row 13 garbage as I couldn't tell what happened with it without this change.

Replay Tests

Fixes #899
In previous versions, non-vs games wouldn't log achieved combos in the stack's combos table. The new checkMatches fixes this and I have adjusted the combo values in the respective asserts.