Fixes to problems found in beta
Make tests assert early if they game over before they are supposed to.
set chaining to nil instead of false to be consistent
We need to mark all panels as state changed in case they need to match for clear puzzles / active puzzles.
We also need to not update panels during updatePanels if the countdown hasn't finished so we don't clear the stateChanged flag (or do unnecessary work)
finally delete an unused variable