matchMetalAndGarbageClearsAllMetalTest
When a match touches metal and a garbage block that also touches the metal we were not clearing the full metal.

The problem was if we processed the non metal clear after the metal clear we would clear the matches metal flag.

I think a similar thing could happen with non metal in the opposite order.

In any case we should be ORing in the flags instead so they can be combined.

Added test and made sure it worked in replay.