Fix fall propagation with garbage
Fall propagation was not happening in the non normal 0 color case.
Wrote a unit test and verified it was fixed by moving the if case out of the inner if statement which was I think only intended for the other case.
All unit tests pass.
Also fixed a rollback problem while debugging in replay mode. Panels in 13 and 14 row might not exist in the destination on rollback, in those limited cases we need to make a panel object still.
Changed the factory method to be able to not increment the ID.