Fix absolute multibar coordinates
There are two problems with the multibar:
1. The coordinates aren't relative to the stack and in absolute coordinates in v3 themes like they should be
2. There was problems with either absolute or non absolute multibar in beta, I don't remember which.

This PR was an attempt to fix both up but I never finished the work.
I was trying to fix the two problems above, and in the process I tried to cleanup the multibar code so it wasn't duplicated for absolute and non absolute and also properly used quads for the absolute multibar so that you could have a gradient or some other non solid fill for the bars.

Unfortunately I never finished it. Posting this in case someone else wants to take it over.