Fix quit button in main menu
The return value for the quit conditional inside CustomRun.processEvents() is intended for the function CustomRun.innerRun().  It became impossible to quit the game without terminating it or using a task manager.

I added a local should_quit boolean which is set to true if the user selects quit, and is returned after processing the events instead of during.