Review feedback
Ironically the coordinates of telegraph where correct and adding 1 to combo looked the same even though chain was actually the one that went up 1. Fixed that and also documented the variables to be only attack draw position. Since push pushes the combo and chain seperately its not clean to do the work in telegraph, but seems fine to do in check matches for now since its also done for cards, at least we are explicit about the variable names now.