Menu Refactor 2
All menu items are now expected to be created with Menu.createMenuItem
This is a bottleneck we can use later to make a concrete menu item if needed.
Also removed unneeded wrapping of arrays as we don't use those anymore
Options menu now trusts layout instead of trying to force stuff itself
Call draw on uiroot so we don't lose a transform
Setup button group to have a proper overall width and height
make label have a top default valign so it doesn't fight the menu. Every label thats in a button gets centered anyways
Fixes in Menu.lua plus helper constructors to have a bottleneck
allow passing in items to StackPanel
Don't hide/show all children when the parent is changed, that was messing up the stepper and isn't needed as the parent hides all children when hidden